Group data columns by block root instead of column index in database (#7199)
For `deneb.BlobSidecar`, #4675 introduced a `blobkey` function that
maximally spreads blob data that pertains to a single block across the
database. This design prevents efficiently deleting data per block root.
Further, the `blobkey` function is endianness dependent, so a database
cannot be migrated between little-endian and big-endian systems.

#6660 then seems to have copied those flaws in `columnkey`, but because
that feature is only used from Fulu and later, it can still be fixed.
Do so now, before there is an expectation of a reliable Fulu DB scheme.