Rework process shutdown, simplified (fixes #3235) (#7418)
https://github.com/status-im/nimbus-eth2/pull/7404 attempted to use
`waitSignal` to handle shutdown notifications but this turned out to be
unreliable, in particular due to multi-threading and cross-platform
differences.

This PR fixes the crash in signal handlers by removing logging but
otherwise retains the polling nature of the shutdown initiation - it
also allows nimbus eth1/2 to share the same cross-thread mechanism for
initiating a shutdown.