Update data directory defaults to include namespace, network (#7422)
* Update data directory defaults to include namespace, network

Running `nimbus_xxx` without `--data-dir` currently results in the data
directories colliding, for several reaons:

* Beacon node, validator client and signing node all use the same
validator folders derived from the data directory - this means that when
running either of those processes together based on defaults, the
processes will attempt to load the same keys which is a bad idea -
separating the folders by namespace gives an extra layer of safety
* Running with `--network` and without `--data-dir` fails due to
incompatible databases - adding the network name helps keep each network
separate

This change only affects the defaults - an explicit `--data-dir`
continues to work as it did before. Also, if the legacy folder exists,
it will be used in preference over the new structure.

While we're here, we also take the opportunity to:

* use `LocalAppData` on Windows, to avoid roaming the blockchain
database
* use `XDG_STATE_HOME` on linux, in line with XDG spec 0.8 (2021)
* rework `nimbus_binary_common` to reduce eth2-specific quirks such that
it can be reused in non-eth2 binaries
* clean up after VC stopped using `runSlotLoop`
* improve compatibilty for out-of-git builds when fetching git revision

* lint

* split

* export

* rm help texts

they don't get expanded

* special-case some network names (minimal in particular needs this)

* what a mess