Store runtime SECONDS_PER_SLOT value in RuntimeConfig (#7479)
As an initial step for making SECONDS_PER_SLOT configurable, store the
value from config.yaml in RuntimeConfig. As the rest of the code still
uses the compile time constant from the {.intdefine.}, the preset loader
still checks the value to be compatible (effectively keeping it static):

> Cannot override config (required: SECONDS_PER_SLOT==12 - config: SECONDS_PER_SLOT=5)

Storing the parsed value is still necessary eventually, doing it now to
keep the rest of the diffs small.