Reply with Access-Control-Expose-Headers if configured with CORS (#7473)
* Reply with Access-Control-Expose-Headers if configured with CORS

When CORS is used, i.e., when the Nimbus beacon API is accessed from
a web browser, by default, custom headers are not accessible from the
client side JavaScript code, as they might contain sensitive info.

Because beacon-API uses custom HTTP headers, we have to send a second
header to actually make them available. This PR does that for:

- Eth-Consensus-Version
- Eth-Execution-Payload-Blinded
- Eth-Execution-Payload-Value
- Eth-Consensus-Block-Value

The Access-Control-Expose-Headers header is sent if Nimbus is configured
with --rest-allow-origin, matching the trigger in Presto that sends the
Access-Control-Allow-Origin header.

* Lint