Store runtime SECONDS_PER_SLOT value in ValidatorClient (#7482)
* Store runtime SECONDS_PER_SLOT value in ValidatorClient

To make SECONDS_PER_SLOT configurable, the validator client needs to be
able to parse it from the config endpoint and use it for validating
compatibility across different beacon nodes.

- When genesis time is initialized, also fetch and store the time config
- When checking compatibility, compare against the local time config

As the rest of the code is still based on the preset constant, it is
also ensured that any non-standard time configs get rejected, same as
currently done, just a bit later in the logic.

* Format error message on out of range value

* SECONDS_PER_SLOT should be taken after genesis is determined

* Fetch time config after genesis is initialized

* Clean debug line

* timeCfg -> timeConfig

* "time config" -> "time configuration settings"

* "unsuccessful" -> "unexpected" for non-200 code

* Log phrasing

* Remove duplicate log

* Move seconds per slot param to back

* Rename time -> getTimeConfig, and constants to caps lock style